requirements.txt
pyspark==3.5.0
redis 
psycopg2-binary
kafka-python
numpy
pandas==1.3.5
scikit-learn==1.3.2
joblib==1.3.2
lightgbm

# dash
# dash-bootstrap-components
# plotly
# psycopg2-binary
